LPS-114731 Separate into interface and base class
