+++ modules/apps/info/info-api/src/main/java/com/liferay/info/item/InfoItemIdentifier.java
@@ -22,21 +22,15 @@ import org.osgi.annotation.versioning.ProviderType;
  * @author Jürgen Kappler
  */
 @ProviderType
-public abstract class InfoItemIdentifier {
+public interface InfoItemIdentifier {
 
 	public static final String VERSION_LATEST = "VERSION_LATEST";
 
 	public static final String VERSION_LATEST_APPROVED =
 		"VERSION_LATEST_APPROVED";
 
-	public Optional<String> getVersionOptional() {
-		return Optional.ofNullable(_version);
-	}
+	public Optional<String> getVersionOptional();
 
-	public void setVersion(String version) {
-		_version = version;
-	}
-
-	private String _version;
+	public void setVersion(String version);
 
 }
